✨ feat: add DynamicPreview component and refactor Hero component to utilize it; rename Preview component file for better organization
