+++ app/components/sections/Hero.tsx
@@ -1,7 +1,7 @@
 "use client";
 // import { Button } from "../Button";
-import Image from "next/image";
-import github from "@/public/github.svg";
+
+import { DynamicPreview } from "../previews/DynamicPreview";
 export default function Hero() {
   return (
     <section
@@ -46,31 +46,7 @@ export default function Hero() {
           </div>
         </div>
         {/* preview */}
-        <div className="w-full sm:max-w-full md:max-w-screen-md lg:max-w-screen-lg backdrop-blur-sm bg-white/10 dark:bg-black/20 relative flex items-center justify-center rounded-xl aspect-[3/1] overflow-hidden border border-white/20 p-8 shadow-xl flex-col font-[montserrat] text-white">
-          {/* Radial gradient overlay with softer transition */}
-          <div className="absolute pointer-events-none inset-0 flex items-center justify-center [mask-image:radial-gradient(ellipse_at_center,transparent_40%,rgba(1,30,160,10))]">
-            {/* Dot pattern with reduced opacity */}
-            <div className="absolute inset-0 bg-dot-black"></div>
-            <div className="absolute inset-0 bg-grid-small-main-blue/40"></div>
-          </div>
-
-          <p className="lg:text-4xl font-bold relative z-20 bg-clip-text text-transparent bg-gradient-to-b from-white to-white/70 py-4">
-            Web Developer
-          </p>
-          <p className="text-white/90 relative z-20 text-xs mb-5">
-            I like to create things on the web
-          </p>
-
-          <div className="  flex items-center gap-2  backdrop-blur-md px-4 py-1 rounded-full bg-black/20">
-            <Image src={github} alt="github" className="text-white h-5 w-5" />
-            <a
-              href="https://github.com/introvertio/better-banners/"
-              className="text-white text-xs font-medium"
-            >
-              Open Source
-            </a>
-          </div>
-        </div>
+        <DynamicPreview/>
 
         {/* Feature badges */}
         <div className=" flex gap-4 my-3 justify-center">
