+++ app/components/sections/Hero2.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const Hero2 = () => {
+  return (
+    <div>Hero2</div>
+  )
+}
